Reuse the existing authenticators with the new implementation
instead of keeping two implementations for the authentication,
consolidate them into one implementation. For the legacy authenticators
configuration, adapt them to use the new authenticator's provider interface.
