+++ authentication/http.go
@@ -18,16 +18,16 @@ func (c contextKey) String() string {
 }
 
 const (
-	// AccessTokenKey is the key that holds the bearer token in a request context.
-	AccessTokenKey contextKey = "accessToken"
-	// GroupsKey is the key that holds the groups in a request context.
-	GroupsKey contextKey = "groups"
-	// SubjectKey is the key that holds the subject in a request context.
-	SubjectKey contextKey = "subject"
-	// TenantKey is the key that holds the tenant in a request context.
-	TenantKey contextKey = "tenant"
-	// TenantIDKey is the key that holds the tenant ID in a request context.
-	TenantIDKey contextKey = "tenantID"
+	// accessTokenKey is the key that holds the bearer token in a request context.
+	accessTokenKey contextKey = "accessToken"
+	// groupsKey is the key that holds the groups in a request context.
+	groupsKey contextKey = "groups"
+	// subjectKey is the key that holds the subject in a request context.
+	subjectKey contextKey = "subject"
+	// tenantKey is the key that holds the tenant in a request context.
+	tenantKey contextKey = "tenant"
+	// tenantIDKey is the key that holds the tenant ID in a request context.
+	tenantIDKey contextKey = "tenantID"
 )
 
 // WithTenant finds the tenant from the URL parameters and adds it to the request context.
@@ -35,7 +35,7 @@ func WithTenant(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		tenant := chi.URLParam(r, "tenant")
 		next.ServeHTTP(w, r.WithContext(
-			context.WithValue(r.Context(), TenantKey, tenant),
+			context.WithValue(r.Context(), tenantKey, tenant),
 		))
 	})
 }
@@ -47,7 +47,7 @@ func WithTenantID(tenantIDs map[string]string) Middleware {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			tenant := chi.URLParam(r, "tenant")
 			next.ServeHTTP(w, r.WithContext(
-				context.WithValue(r.Context(), TenantIDKey, tenantIDs[tenant]),
+				context.WithValue(r.Context(), tenantIDKey, tenantIDs[tenant]),
 			))
 		})
 	}
@@ -61,7 +61,7 @@ func WithAccessToken() Middleware {
 			rawToken := r.Header.Get("Authorization")
 			token := rawToken[strings.LastIndex(rawToken, " ")+1:]
 			next.ServeHTTP(w, r.WithContext(
-				context.WithValue(r.Context(), AccessTokenKey, token),
+				context.WithValue(r.Context(), accessTokenKey, token),
 			))
 		})
 	}
@@ -80,7 +80,7 @@ func WithTenantHeader(header string, tenantIDs map[string]string) Middleware {
 
 // GetTenant extracts the tenant from provided context.
 func GetTenant(ctx context.Context) (string, bool) {
-	value := ctx.Value(TenantKey)
+	value := ctx.Value(tenantKey)
 	tenant, ok := value.(string)
 
 	return tenant, ok
@@ -88,7 +88,7 @@ func GetTenant(ctx context.Context) (string, bool) {
 
 // GetTenantID extracts the tenant ID from provided context.
 func GetTenantID(ctx context.Context) (string, bool) {
-	value := ctx.Value(TenantIDKey)
+	value := ctx.Value(tenantIDKey)
 	id, ok := value.(string)
 
 	return id, ok
@@ -96,7 +96,7 @@ func GetTenantID(ctx context.Context) (string, bool) {
 
 // GetSubject extracts the subject from provided context.
 func GetSubject(ctx context.Context) (string, bool) {
-	value := ctx.Value(SubjectKey)
+	value := ctx.Value(subjectKey)
 	subject, ok := value.(string)
 
 	return subject, ok
@@ -104,7 +104,7 @@ func GetSubject(ctx context.Context) (string, bool) {
 
 // GetGroups extracts the groups from provided context.
 func GetGroups(ctx context.Context) ([]string, bool) {
-	value := ctx.Value(GroupsKey)
+	value := ctx.Value(groupsKey)
 	groups, ok := value.([]string)
 
 	return groups, ok
@@ -112,7 +112,7 @@ func GetGroups(ctx context.Context) ([]string, bool) {
 
 // GetAccessToken extracts the access token from the provided context.
 func GetAccessToken(ctx context.Context) (string, bool) {
-	value := ctx.Value(AccessTokenKey)
+	value := ctx.Value(accessTokenKey)
 	token, ok := value.(string)
 
 	return token, ok
