+++ authentication/metrics.go
@@ -5,11 +5,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promauto"
 )
 
-const (
-	providerOIDC      = "oidc"
-	providerOpenShift = "openshift"
-)
-
 func RegisterTenantsFailingMetric(reg prometheus.Registerer) *prometheus.CounterVec {
 	return promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
 		Namespace: "observatorium",
