+++ authentication/oidc.go
@@ -4,12 +4,13 @@ import (
 	"context"
 	"crypto/tls"
 	"crypto/x509"
+	"encoding/pem"
 	"fmt"
+	"io/ioutil"
 	"net"
 	"net/http"
 	"path"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/coreos/go-oidc"
@@ -17,150 +18,92 @@ import (
 	"github.com/go-chi/chi"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/mitchellh/mapstructure"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"golang.org/x/oauth2"
 )
 
-// TenantOIDCConfig represents the OIDC configuration for a single tenant.
-type TenantOIDCConfig struct {
-	Tenant string
+// OIDCAuthenticatorType represents the oidc authentication provider type.
+const OIDCAuthenticatorType = "oidc"
 
-	OIDCConfig
+//nolint:gochecknoinits
+func init() {
+	// authenticatorFactories[OIDCAuthenticatorType] = newOIDCAuthenticator
+	onboardNewProvider(OIDCAuthenticatorType, newOIDCAuthenticator)
 }
 
-// OIDCConfig represents a generic OIDC configuration.
-type OIDCConfig struct {
-	IssuerCA      *x509.Certificate
-	IssuerURL     string
-	ClientID      string
-	ClientSecret  string
-	GroupClaim    string
-	RedirectURL   string
-	UsernameClaim string
+// oidcConfig represents the oidc authenticator config.
+type oidcConfig struct {
+	ClientID      string `json:"clientID"`
+	ClientSecret  string `json:"clientSecret"`
+	GroupClaim    string `json:"groupClaim"`
+	IssuerRawCA   []byte `json:"issuerCA"`
+	IssuerCAPath  string `json:"issuerCAPath"`
+	issuerCA      *x509.Certificate
+	IssuerURL     string `json:"issuerURL"`
+	RedirectURL   string `json:"redirectURL"`
+	UsernameClaim string `json:"usernameClaim"`
 }
 
-// OIDCHandlers holds handlers and a set of middlewares for all
-// tenants that is able to authenticate requests and provide the
-// authorization code grant flow for users.
-type OIDCHandlers struct {
-	mtx         sync.RWMutex
-	handlers    map[string]http.Handler
-	middlewares map[string]Middleware
-
-	logger     log.Logger
-	retryCount *prometheus.CounterVec
-}
-
-// NewOIDCHandlers instantiates OIDC handlers.
-func NewOIDCHandlers(l log.Logger, retryCount *prometheus.CounterVec) *OIDCHandlers {
-	return &OIDCHandlers{
-		handlers:    make(map[string]http.Handler),
-		middlewares: make(map[string]Middleware),
-		logger:      l,
-		retryCount:  retryCount,
-	}
-}
-
-// AddOIDCForTenant adds a new handler and middleware for a new tenant, based on
-// the tenant OIDC configuration. If a connection with the OIDC provider
-// cannot be established, the method will retry to connect with an increasing
-// backoff.
-func (oh *OIDCHandlers) AddOIDCForTenant(prefix string, config TenantOIDCConfig) {
-	const (
-		loginRoute    = "/login"
-		callbackRoute = "/callback"
-	)
-
-	ctx := context.Background()
-	b := backoff.New(ctx, backoff.Config{
-		Min:        500 * time.Millisecond,
-		Max:        5 * time.Second,
-		MaxRetries: 0, // Retry indefinitely.
-	})
-
-	go func() {
-		r := chi.NewRouter()
-
-		for b.Reset(); b.Ongoing(); {
-			p, err := NewProvider(ctx, oh.logger, getCookieForTenant(config.Tenant), "/"+config.Tenant, config.OIDCConfig)
-			if err != nil {
-				level.Warn(oh.logger).Log("msg", "failed to instantiate OIDC provider for tenant", "tenant", config.Tenant, "error", err)
-				oh.retryCount.WithLabelValues(config.Tenant, providerOIDC).Inc()
-				b.Wait()
-				continue
-			}
-
-			r.Handle(loginRoute, otelhttp.WithRouteTag(prefix+loginRoute, p.LoginHandler()))
-			r.Handle(callbackRoute, otelhttp.WithRouteTag(prefix+callbackRoute, p.CallbackHandler()))
-
-			oh.mtx.Lock()
-			oh.handlers[config.Tenant] = r
-			oh.middlewares[config.Tenant] = p.Middleware()
-			oh.mtx.Unlock()
-
-			level.Info(oh.logger).Log("msg", "OIDC provider instantiated for tenant", "tenant", config.Tenant)
-
-			return
-		}
-	}()
-}
-
-// GetTenantMiddleware an OIDC middleware for the specified tenant, if found.
-func (oh *OIDCHandlers) GetTenantMiddleware(tenant string) (Middleware, bool) {
-	oh.mtx.RLock()
-	mw, ok := oh.middlewares[tenant]
-	oh.mtx.RUnlock()
-
-	return mw, ok
-}
-
-// Router returns a router with handlers for all registered tenants.
-func (oh *OIDCHandlers) Router() *chi.Mux {
-	r := chi.NewRouter()
-	r.Mount("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		tenant, ok := GetTenant(r.Context())
-		const msg = "error finding tenant"
-		if !ok {
-			level.Warn(oh.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		oh.mtx.RLock()
-		h, ok := oh.handlers[tenant]
-		oh.mtx.RUnlock()
-		if !ok {
-			level.Debug(oh.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusUnauthorized)
-			return
-		}
-		h.ServeHTTP(w, r)
-	}))
-
-	return r
-}
-
-func getCookieForTenant(tenant string) string {
-	return fmt.Sprintf("observatorium_%s", tenant)
-}
-
-// OIDCProvider wraps the oidc.Provider in an opinionated way to create login
-// and callback handlers as well as HTTP middlewares to authenticate requests
-// against the provider.
-type OIDCProvider struct {
+type oidcAuthenticator struct {
+	tenant       string
 	logger       log.Logger
+	config       oidcConfig
 	provider     *oidc.Provider
+	verifier     *oidc.IDTokenVerifier
 	client       *http.Client
-	config       OIDCConfig
 	cookieName   string
 	redirectURL  string
 	oauth2Config oauth2.Config
-	verifier     *oidc.IDTokenVerifier
+	handler      http.Handler
 }
 
-// NewProvider creates a new OIDCProvider.
-func NewProvider(ctx context.Context, logger log.Logger, cookieName, redirectURL string, config OIDCConfig) (*OIDCProvider, error) {
+func newOIDCAuthenticator(c map[string]interface{}, tenant string,
+	registrationRetryCount *prometheus.CounterVec, logger log.Logger) (Provider, error) {
+	var config oidcConfig
+
+	const (
+		loginRoute    = "/login"
+		callbackRoute = "/callback"
+		handlerPrefix = "/oidc/{tenant}"
+	)
+
+	ctx := context.Background()
+
+	err := mapstructure.Decode(c, &config)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(config.IssuerURL) == 0 {
+		return nil, fmt.Errorf("issuerURL is required")
+	}
+
+	if config.IssuerCAPath != "" {
+		IssuerRawCA, err := ioutil.ReadFile(config.IssuerCAPath)
+		if err != nil {
+			return nil, fmt.Errorf("failed to read issuer ca file: %s", err.Error())
+		}
+
+		config.IssuerRawCA = IssuerRawCA
+	}
+
+	if len(config.IssuerRawCA) != 0 {
+		block, _ := pem.Decode(config.IssuerRawCA)
+		if block == nil {
+			return nil, fmt.Errorf("failed to parse issuer CA certificate PEM")
+		}
+
+		cert, err := x509.ParseCertificate(block.Bytes)
+
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse issuer certificate: %s", err.Error())
+		}
+
+		config.issuerCA = cert
+	}
+
 	t := &http.Transport{
 		Proxy: http.ProxyFromEnvironment,
 		DialContext: (&net.Dialer{
@@ -173,11 +116,12 @@ func NewProvider(ctx context.Context, logger log.Logger, cookieName, redirectURL
 		TLSHandshakeTimeout:   10 * time.Second,
 		ExpectContinueTimeout: 1 * time.Second,
 	}
-	if config.IssuerCA != nil {
+
+	if config.issuerCA != nil {
 		t.TLSClientConfig = &tls.Config{
 			RootCAs: x509.NewCertPool(),
 		}
-		t.TLSClientConfig.RootCAs.AddCert(config.IssuerCA)
+		t.TLSClientConfig.RootCAs.AddCert(config.issuerCA)
 	}
 
 	client := &http.Client{
@@ -185,10 +129,7 @@ func NewProvider(ctx context.Context, logger log.Logger, cookieName, redirectURL
 		Transport: t,
 	}
 
-	provider, err := oidc.NewProvider(oidc.ClientContext(ctx, client), config.IssuerURL)
-	if err != nil {
-		return nil, err
-	}
+	provider := newOIDCProvider(oidc.ClientContext(ctx, client), tenant, client, config.IssuerURL, registrationRetryCount, logger)
 
 	oauth2Config := oauth2.Config{
 		ClientID:     config.ClientID,
@@ -200,28 +141,137 @@ func NewProvider(ctx context.Context, logger log.Logger, cookieName, redirectURL
 
 	verifier := provider.Verifier(&oidc.Config{ClientID: config.ClientID})
 
-	return &OIDCProvider{
+	oidcProvider := &oidcAuthenticator{
+		tenant:       tenant,
 		logger:       logger,
-		provider:     provider,
-		client:       client,
 		config:       config,
-		cookieName:   cookieName,
-		redirectURL:  redirectURL,
 		oauth2Config: oauth2Config,
+		provider:     provider,
 		verifier:     verifier,
-	}, nil
+		client:       client,
+		cookieName:   fmt.Sprintf("observatorium_%s", tenant),
+	}
+
+	r := chi.NewRouter()
+	r.Handle(loginRoute, otelhttp.WithRouteTag(handlerPrefix+loginRoute, oidcProvider.oidcLoginHandler(&oauth2Config)))
+	r.Handle(callbackRoute, otelhttp.WithRouteTag(handlerPrefix+callbackRoute, oidcProvider.oidcCallBackHandler()))
+	oidcProvider.handler = r
+
+	return oidcProvider, nil
 }
 
-// OIDCProvider returns the internal oidc.Provider, this can be used to for
-// example access the UserInfo function of the provider.
-func (p *OIDCProvider) OIDCProvider() *oidc.Provider {
-	return p.provider
+func newOIDCProvider(ctx context.Context, tenant string, client *http.Client, issuerURL string,
+	registrationRetryCount *prometheus.CounterVec, logger log.Logger) *oidc.Provider {
+	var provider *oidc.Provider
+
+	var err error
+
+	b := backoff.New(ctx, backoff.Config{
+		Min:        500 * time.Millisecond,
+		Max:        5 * time.Second,
+		MaxRetries: 0, // Retry indefinitely.
+	})
+
+	for b.Reset(); b.Ongoing(); {
+		provider, err = oidc.NewProvider(oidc.ClientContext(ctx, client), issuerURL)
+		if err != nil {
+			level.Error(logger).Log(
+				"tenant", tenant,
+				"msg", fmt.Sprintf("failed to initialize authenticator %s after %d retries: %s",
+					OIDCAuthenticatorType, b.NumRetries(), err))
+			registrationRetryCount.WithLabelValues(tenant, OIDCAuthenticatorType).Inc()
+			b.Wait()
+
+			continue
+		}
+
+		break
+	}
+
+	return provider
 }
 
-// Middleware creates an HTTP middleware, that can be used to wrap a
-// http.Handler to authenticate requests for it.
+func (a oidcAuthenticator) oidcLoginHandler(oauth2Config *oauth2.Config) http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		url := oauth2Config.AuthCodeURL(state)
+		http.Redirect(w, r, url, http.StatusSeeOther)
+	})
+}
+
+func (a oidcAuthenticator) oidcCallBackHandler() http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		ctx := oidc.ClientContext(r.Context(), a.client)
+
+		if errMsg := r.URL.Query().Get("error"); errMsg != "" {
+			desc := r.URL.Query().Get("error_description")
+			msg := fmt.Sprintf("%s: %s", errMsg, desc)
+			level.Debug(a.logger).Log("msg", msg)
+			http.Error(w, msg, http.StatusBadRequest)
+			return
+		}
+
+		queryCode := r.URL.Query().Get("code")
+		if queryCode == "" {
+			const msg = "no code in request"
+			level.Debug(a.logger).Log("msg", msg)
+			http.Error(w, msg, http.StatusBadRequest)
+			return
+		}
+		queryState := r.URL.Query().Get("state")
+		if queryState != state {
+			const msg = "incorrect state in request"
+			level.Debug(a.logger).Log("msg", msg)
+			http.Error(w, msg, http.StatusBadRequest)
+			return
+		}
+
+		token, err := a.oauth2Config.Exchange(ctx, queryCode)
+		if err != nil {
+			msg := fmt.Sprintf("failed to get token: %v", err)
+			level.Warn(a.logger).Log("msg", msg, "err", err)
+			http.Error(w, msg, http.StatusInternalServerError)
+			return
+		}
+
+		rawIDToken, ok := token.Extra("id_token").(string)
+		if !ok {
+			const msg = "no id_token in token response"
+			level.Warn(a.logger).Log("msg", msg)
+			http.Error(w, msg, http.StatusInternalServerError)
+			return
+		}
+
+		_, err = a.verifier.Verify(ctx, rawIDToken)
+		if err != nil {
+			msg := fmt.Sprintf("failed to verify ID token: %v", err)
+			level.Warn(a.logger).Log("msg", msg)
+			http.Error(w, msg, http.StatusInternalServerError)
+			return
+		}
+
+		http.SetCookie(w, &http.Cookie{
+			Name:    a.cookieName,
+			Value:   rawIDToken,
+			Path:    "/",
+			Expires: token.Expiry,
+		})
+
+		http.Redirect(w, r, a.redirectURL, http.StatusFound)
+	})
+}
+
+func (a oidcAuthenticator) Handler() (string, http.Handler) {
+	r := chi.NewRouter()
+
+	r.Mount("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		a.handler.ServeHTTP(w, r)
+	}))
+
+	return "/oidc/{tenant}", r
+}
+
 //nolint:gocognit
-func (p *OIDCProvider) Middleware() Middleware {
+func (a oidcAuthenticator) Middleware() Middleware {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			var token string
@@ -231,19 +281,19 @@ func (p *OIDCProvider) Middleware() Middleware {
 				authorization := strings.Split(authorizationHeader, " ")
 				if len(authorization) != 2 {
 					const msg = "invalid Authorization header"
-					level.Debug(p.logger).Log("msg", msg)
+					level.Debug(a.logger).Log("msg", msg)
 					http.Error(w, msg, http.StatusUnauthorized)
 					return
 				}
 
 				token = authorization[1]
 			} else {
-				cookie, err := r.Cookie(p.cookieName)
+				cookie, err := r.Cookie(a.cookieName)
 				if err != nil {
 					tenant, ok := GetTenant(r.Context())
 					if !ok {
 						const msg = "error finding tenant"
-						level.Warn(p.logger).Log("msg", msg)
+						level.Warn(a.logger).Log("msg", msg)
 						http.Error(w, msg, http.StatusInternalServerError)
 						return
 					}
@@ -255,54 +305,54 @@ func (p *OIDCProvider) Middleware() Middleware {
 				token = cookie.Value
 			}
 
-			idToken, err := p.verifier.Verify(oidc.ClientContext(r.Context(), p.client), token)
+			idToken, err := a.verifier.Verify(oidc.ClientContext(r.Context(), a.client), token)
 			if err != nil {
 				const msg = "failed to authenticate"
-				level.Debug(p.logger).Log("msg", msg, "err", err)
+				level.Debug(a.logger).Log("msg", msg, "err", err)
 				http.Error(w, msg, http.StatusBadRequest)
 				return
 			}
 
 			sub := idToken.Subject
-			if p.config.UsernameClaim != "" {
+			if a.config.UsernameClaim != "" {
 				claims := map[string]interface{}{}
 				if err := idToken.Claims(&claims); err != nil {
 					const msg = "failed to read claims"
-					level.Warn(p.logger).Log("msg", msg, "err", err)
+					level.Warn(a.logger).Log("msg", msg, "err", err)
 					http.Error(w, msg, http.StatusInternalServerError)
 					return
 				}
-				rawUsername, ok := claims[p.config.UsernameClaim]
+				rawUsername, ok := claims[a.config.UsernameClaim]
 				if !ok {
 					const msg = "username cannot be empty"
-					level.Debug(p.logger).Log("msg", msg)
+					level.Debug(a.logger).Log("msg", msg)
 					http.Error(w, msg, http.StatusBadRequest)
 					return
 				}
 				username, ok := rawUsername.(string)
 				if !ok || username == "" {
 					const msg = "invalid username claim value"
-					level.Debug(p.logger).Log("msg", msg)
+					level.Debug(a.logger).Log("msg", msg)
 					http.Error(w, msg, http.StatusBadRequest)
 					return
 				}
 				sub = username
 			}
-			ctx := context.WithValue(r.Context(), SubjectKey, sub)
+			ctx := context.WithValue(r.Context(), subjectKey, sub)
 
-			if p.config.GroupClaim != "" {
+			if a.config.GroupClaim != "" {
 				var groups []string
 				claims := map[string]interface{}{}
 				if err := idToken.Claims(&claims); err != nil {
 					const msg = "failed to read claims"
-					level.Warn(p.logger).Log("msg", msg, "err", err)
+					level.Warn(a.logger).Log("msg", msg, "err", err)
 					http.Error(w, msg, http.StatusInternalServerError)
 					return
 				}
-				rawGroup, ok := claims[p.config.GroupClaim]
+				rawGroup, ok := claims[a.config.GroupClaim]
 				if !ok {
 					const msg = "group cannot be empty"
-					level.Debug(p.logger).Log("msg", msg)
+					level.Debug(a.logger).Log("msg", msg)
 					http.Error(w, msg, http.StatusBadRequest)
 					return
 				}
@@ -317,82 +367,10 @@ func (p *OIDCProvider) Middleware() Middleware {
 						groups = append(groups, fmt.Sprintf("%v", v[i]))
 					}
 				}
-				ctx = context.WithValue(ctx, GroupsKey, groups)
+				ctx = context.WithValue(ctx, groupsKey, groups)
 			}
 
 			next.ServeHTTP(w, r.WithContext(ctx))
 		})
 	}
-}
-
-// LoginHandler returns a http.Handler, that starts the OIDC flow.
-func (p *OIDCProvider) LoginHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		url := p.oauth2Config.AuthCodeURL(state)
-		http.Redirect(w, r, url, http.StatusSeeOther)
-	})
-}
-
-// CallbackHandler returns a http.Handler, that is meant to be redirected to
-// after the provider has authenticated the user.
-func (p *OIDCProvider) CallbackHandler() http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := oidc.ClientContext(r.Context(), p.client)
-
-		if errMsg := r.URL.Query().Get("error"); errMsg != "" {
-			desc := r.URL.Query().Get("error_description")
-			msg := fmt.Sprintf("%s: %s", errMsg, desc)
-			level.Debug(p.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-
-		queryCode := r.URL.Query().Get("code")
-		if queryCode == "" {
-			const msg = "no code in request"
-			level.Debug(p.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-		queryState := r.URL.Query().Get("state")
-		if queryState != state {
-			const msg = "incorrect state in request"
-			level.Debug(p.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-
-		token, err := p.oauth2Config.Exchange(ctx, queryCode)
-		if err != nil {
-			msg := fmt.Sprintf("failed to get token: %v", err)
-			level.Warn(p.logger).Log("msg", msg, "err", err)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		rawIDToken, ok := token.Extra("id_token").(string)
-		if !ok {
-			const msg = "no id_token in token response"
-			level.Warn(p.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		_, err = p.verifier.Verify(ctx, rawIDToken)
-		if err != nil {
-			msg := fmt.Sprintf("failed to verify ID token: %v", err)
-			level.Warn(p.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		http.SetCookie(w, &http.Cookie{
-			Name:    p.cookieName,
-			Value:   rawIDToken,
-			Path:    "/",
-			Expires: token.Expiry,
-		})
-
-		http.Redirect(w, r, p.redirectURL, http.StatusFound)
-	})
 }
