--- authentication/providers/oidc.go
@@ -1,376 +1,0 @@
-package providers
-
-import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"encoding/pem"
-	"fmt"
-	"io/ioutil"
-	"net"
-	"net/http"
-	"path"
-	"strings"
-	"time"
-
-	"github.com/coreos/go-oidc"
-	"github.com/efficientgo/tools/core/pkg/backoff"
-	"github.com/go-chi/chi"
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/log/level"
-	"github.com/mitchellh/mapstructure"
-	"github.com/observatorium/api/authentication"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-	"golang.org/x/oauth2"
-)
-
-const (
-	state                 = "I love Observatorium"
-	oidcAuthenticatorType = "oidc"
-)
-
-//nolint:gochecknoinits
-func init() {
-	authenticatorsFactories[oidcAuthenticatorType] = newOIDCAuthenticator
-}
-
-type OIDCConfig struct {
-	ClientID      string `json:"clientID"`
-	ClientSecret  string `json:"clientSecret"`
-	GroupClaim    string `json:"groupClaim"`
-	IssuerRawCA   []byte `json:"issuerCA"`
-	IssuerCAPath  string `json:"issuerCAPath"`
-	issuerCA      *x509.Certificate
-	IssuerURL     string `json:"issuerURL"`
-	RedirectURL   string `json:"redirectURL"`
-	UsernameClaim string `json:"usernameClaim"`
-}
-
-type OIDCAuthenticator struct {
-	tenant       string
-	logger       log.Logger
-	Config       OIDCConfig
-	provider     *oidc.Provider
-	verifier     *oidc.IDTokenVerifier
-	client       *http.Client
-	cookieName   string
-	redirectURL  string
-	oauth2Config oauth2.Config
-	Handler      http.Handler
-}
-
-func newOIDCAuthenticator(c map[string]interface{}, tenant string, logger log.Logger) (AuthenticationProvider, error) {
-	var config OIDCConfig
-
-	const (
-		loginRoute    = "/login"
-		callbackRoute = "/callback"
-		handlerPrefix = "/oidc/{tenant}"
-	)
-
-	ctx := context.Background()
-
-	err := mapstructure.Decode(c, &config)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(config.IssuerURL) == 0 {
-		return nil, fmt.Errorf("issuerURL is required")
-	}
-
-	if config.IssuerCAPath != "" {
-		IssuerRawCA, err := ioutil.ReadFile(config.IssuerCAPath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read issuer ca file: %s", err.Error())
-		}
-
-		config.IssuerRawCA = IssuerRawCA
-	}
-
-	if len(config.IssuerRawCA) != 0 {
-		block, _ := pem.Decode(config.IssuerRawCA)
-		if block == nil {
-			return nil, fmt.Errorf("failed to parse issuer CA certificate PEM")
-		}
-
-		cert, err := x509.ParseCertificate(block.Bytes)
-
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse issuer certificate: %s", err.Error())
-		}
-
-		config.issuerCA = cert
-	}
-
-	t := &http.Transport{
-		Proxy: http.ProxyFromEnvironment,
-		DialContext: (&net.Dialer{
-			Timeout:   30 * time.Second,
-			KeepAlive: 30 * time.Second,
-		}).DialContext,
-		ForceAttemptHTTP2:     true,
-		MaxIdleConns:          100,
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-
-	if config.issuerCA != nil {
-		t.TLSClientConfig = &tls.Config{
-			RootCAs: x509.NewCertPool(),
-		}
-		t.TLSClientConfig.RootCAs.AddCert(config.issuerCA)
-	}
-
-	client := &http.Client{
-		Timeout:   5 * time.Second,
-		Transport: t,
-	}
-
-	provider := newOIDCProvider(oidc.ClientContext(ctx, client), tenant, client, config.IssuerURL, logger)
-
-	oauth2Config := oauth2.Config{
-		ClientID:     config.ClientID,
-		ClientSecret: config.ClientSecret,
-		Endpoint:     provider.Endpoint(),
-		RedirectURL:  config.RedirectURL,
-		Scopes:       []string{"openid", "profile", "email", "groups"},
-	}
-
-	verifier := provider.Verifier(&oidc.Config{ClientID: config.ClientID})
-
-	oidcProvider := &OIDCAuthenticator{
-		tenant:       tenant,
-		logger:       logger,
-		Config:       config,
-		oauth2Config: oauth2Config,
-		provider:     provider,
-		verifier:     verifier,
-		client:       client,
-		cookieName:   fmt.Sprintf("observatorium_%s", tenant),
-	}
-
-	r := chi.NewRouter()
-	r.Handle(loginRoute, otelhttp.WithRouteTag(handlerPrefix+loginRoute, oidcProvider.oidcLoginHandler(&oauth2Config)))
-	r.Handle(callbackRoute, otelhttp.WithRouteTag(handlerPrefix+callbackRoute, oidcProvider.oidcCallBackHandler()))
-	oidcProvider.Handler = r
-
-	return oidcProvider, nil
-}
-
-func newOIDCProvider(ctx context.Context, tenant string, client *http.Client, issuerURL string, logger log.Logger) *oidc.Provider {
-	var provider *oidc.Provider
-
-	var err error
-
-	b := backoff.New(ctx, backoff.Config{
-		Min:        500 * time.Millisecond,
-		Max:        5 * time.Second,
-		MaxRetries: 0, // Retry indefinitely.
-	})
-
-	for b.Reset(); b.Ongoing(); {
-		provider, err = oidc.NewProvider(oidc.ClientContext(ctx, client), issuerURL)
-		if err != nil {
-			level.Error(logger).Log(
-				"tenant", tenant,
-				"msg", fmt.Sprintf("failed to initialize authenticator %s after %d retries: %s",
-					oidcAuthenticatorType, b.NumRetries(), err))
-			registrationRetryCount.WithLabelValues(tenant, oidcAuthenticatorType).Inc()
-			b.Wait()
-
-			continue
-		}
-
-		break
-	}
-
-	return provider
-}
-
-func (a OIDCAuthenticator) oidcLoginHandler(oauth2Config *oauth2.Config) http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		url := oauth2Config.AuthCodeURL(state)
-		http.Redirect(w, r, url, http.StatusSeeOther)
-	})
-}
-
-func (a OIDCAuthenticator) oidcCallBackHandler() http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx := oidc.ClientContext(r.Context(), a.client)
-
-		if errMsg := r.URL.Query().Get("error"); errMsg != "" {
-			desc := r.URL.Query().Get("error_description")
-			msg := fmt.Sprintf("%s: %s", errMsg, desc)
-			level.Debug(a.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-
-		queryCode := r.URL.Query().Get("code")
-		if queryCode == "" {
-			const msg = "no code in request"
-			level.Debug(a.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-		queryState := r.URL.Query().Get("state")
-		if queryState != state {
-			const msg = "incorrect state in request"
-			level.Debug(a.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-
-		token, err := a.oauth2Config.Exchange(ctx, queryCode)
-		if err != nil {
-			msg := fmt.Sprintf("failed to get token: %v", err)
-			level.Warn(a.logger).Log("msg", msg, "err", err)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		rawIDToken, ok := token.Extra("id_token").(string)
-		if !ok {
-			const msg = "no id_token in token response"
-			level.Warn(a.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		_, err = a.verifier.Verify(ctx, rawIDToken)
-		if err != nil {
-			msg := fmt.Sprintf("failed to verify ID token: %v", err)
-			level.Warn(a.logger).Log("msg", msg)
-			http.Error(w, msg, http.StatusInternalServerError)
-			return
-		}
-
-		http.SetCookie(w, &http.Cookie{
-			Name:    a.cookieName,
-			Value:   rawIDToken,
-			Path:    "/",
-			Expires: token.Expiry,
-		})
-
-		http.Redirect(w, r, a.redirectURL, http.StatusFound)
-	})
-}
-
-func (a OIDCAuthenticator) GetHandler() (string, http.Handler) {
-	r := chi.NewRouter()
-
-	r.Mount("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		a.Handler.ServeHTTP(w, r)
-	}))
-
-	return "/oidc/{tenant}", r
-}
-
-func (a OIDCAuthenticator) GetTenant() string {
-	return a.tenant
-}
-
-//nolint:gocognit
-func (a OIDCAuthenticator) AuthenticationMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var token string
-
-		authorizationHeader := r.Header.Get("Authorization")
-		if authorizationHeader != "" {
-			authorization := strings.Split(authorizationHeader, " ")
-			if len(authorization) != 2 {
-				const msg = "invalid Authorization header"
-				level.Debug(a.logger).Log("msg", msg)
-				http.Error(w, msg, http.StatusUnauthorized)
-				return
-			}
-
-			token = authorization[1]
-		} else {
-			cookie, err := r.Cookie(a.cookieName)
-			if err != nil {
-				tenant, ok := authentication.GetTenant(r.Context())
-				if !ok {
-					const msg = "error finding tenant"
-					level.Warn(a.logger).Log("msg", msg)
-					http.Error(w, msg, http.StatusInternalServerError)
-					return
-				}
-				// Redirect users to the OIDC login
-				w.Header().Set("Location", path.Join("/oidc", tenant, "/login"))
-				http.Error(w, "failed to find token", http.StatusFound)
-				return
-			}
-			token = cookie.Value
-		}
-
-		idToken, err := a.verifier.Verify(oidc.ClientContext(r.Context(), a.client), token)
-		if err != nil {
-			const msg = "failed to authenticate"
-			level.Debug(a.logger).Log("msg", msg, "err", err)
-			http.Error(w, msg, http.StatusBadRequest)
-			return
-		}
-
-		sub := idToken.Subject
-		if a.Config.UsernameClaim != "" {
-			claims := map[string]interface{}{}
-			if err := idToken.Claims(&claims); err != nil {
-				const msg = "failed to read claims"
-				level.Warn(a.logger).Log("msg", msg, "err", err)
-				http.Error(w, msg, http.StatusInternalServerError)
-				return
-			}
-			rawUsername, ok := claims[a.Config.UsernameClaim]
-			if !ok {
-				const msg = "username cannot be empty"
-				level.Debug(a.logger).Log("msg", msg)
-				http.Error(w, msg, http.StatusBadRequest)
-				return
-			}
-			username, ok := rawUsername.(string)
-			if !ok || username == "" {
-				const msg = "invalid username claim value"
-				level.Debug(a.logger).Log("msg", msg)
-				http.Error(w, msg, http.StatusBadRequest)
-				return
-			}
-			sub = username
-		}
-		ctx := context.WithValue(r.Context(), authentication.SubjectKey, sub)
-
-		if a.Config.GroupClaim != "" {
-			var groups []string
-			claims := map[string]interface{}{}
-			if err := idToken.Claims(&claims); err != nil {
-				const msg = "failed to read claims"
-				level.Warn(a.logger).Log("msg", msg, "err", err)
-				http.Error(w, msg, http.StatusInternalServerError)
-				return
-			}
-			rawGroup, ok := claims[a.Config.GroupClaim]
-			if !ok {
-				const msg = "group cannot be empty"
-				level.Debug(a.logger).Log("msg", msg)
-				http.Error(w, msg, http.StatusBadRequest)
-				return
-			}
-			switch v := rawGroup.(type) {
-			case string:
-				groups = append(groups, v)
-			case []string:
-				groups = v
-			case []interface{}:
-				groups = make([]string, 0, len(v))
-				for i := range v {
-					groups = append(groups, fmt.Sprintf("%v", v[i]))
-				}
-			}
-			ctx = context.WithValue(ctx, authentication.GroupsKey, groups)
-		}
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
